update on pr

+++ src/main/java/com/trove/gamifiedtrading/data/dto/ConvertUserEntity.java
@@ -2,7 +2,7 @@ package com.trove.gamifiedtrading.data.dto;
 
 public record ConvertUserEntity(
     String username,
-    Integer rank,
+    Long rank,
     Integer gemCount
 ) {
 }
